% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_proposal.r
\name{submit_proposal}
\alias{submit_proposal}
\title{Submit a GSOC Proposal}
\usage{
submit_proposal()
}
\value{
A logical value: \code{TRUE} if run before or on the deadline,
and \code{FALSE} if run after.
}
\description{
A mock function to ostensibly submit proposals to Google Summer of Code.

The deadline for submissions is: \Sexpr[stage=install, results=text]{GSOC:::get_deadline_date()}.

\Sexpr[stage=render, results=rd]{GSOC:::get_deadline_message()}
}
\examples{
submit_proposal()
}
